Do not show Empty Trash button for empty state

+++ src/administrator/components/com_weblinks/src/View/Weblinks/HtmlView.php
@@ -182,7 +182,7 @@ class HtmlView extends BaseHtmlView
 			}
 		}
 
-		if ($this->state->get('filter.published') == -2 && $canDo->get('core.delete'))
+		if (!$this->isEmptyState && $this->state->get('filter.published') == -2 && $canDo->get('core.delete'))
 		{
 			$toolbar->delete('weblinks.delete')
 				->text('JTOOLBAR_EMPTY_TRASH')
